requirements.txt
alembic==1.8.1
fastapi==0.87.0
uvicorn==0.20.0
SQLAlchemy==1.4.44
dependency-injector==4.40.0
python-dotenv==0.21.0
pydantic~=1.10.2
python-jose==3.3.0
cryptography==38.0.3
